feat: draft move to postgresql DB artists and albums

+++ src/modules/album/album.service.ts
@@ -1,9 +1,15 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import {
+  HttpException,
+  HttpStatus,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateAlbumDto } from './dto/create-album.dto';
 import { UpdateAlbumDto } from './dto/update-album.dto';
 import { InMemoryUserStore } from '../../store/in-memory-user-store';
 import { checkFields, createNewAlbum, validateId404 } from './helpers';
 import { PrismaService } from '../../prisma/prisma.service';
+import { StatusCodes } from 'http-status-codes';
 
 @Injectable()
 export class AlbumService {
@@ -44,21 +50,43 @@ export class AlbumService {
   }
 
   async remove(id: string) {
-    validateId404(id, await this.prisma.album.findMany());
+    // validateId404(id, await this.prisma.album.findMany());
     // this.store.albums = this.store.albums.filter((item) => item.id !== id);
-    return await this.prisma.album.delete({ where: { id } });
+    // try {
+    const isAlbumInFavs = await this.prisma.favorites.findFirst({
+      where: { id },
+    });
+    if (isAlbumInFavs) await this.prisma.favorites.delete({ where: { id } });
+    // console.log(id);
 
-    // this.store.tracks.forEach((track, index) => {
-    //   if (track.albumId === id) this.store.tracks[index].albumId = null;
-    // });
+    const isAlbum = await this.prisma.album.findFirst({ where: { id } });
+    if (isAlbum) return await this.prisma.album.delete({ where: { id } });
+    // return await this.prisma.album.delete({ where: { id } });
+    // return;
+    throw new NotFoundException();
 
-    //
-    // const trackListToClean = await this.prisma.track.findMany({ where:  })
-
-    //
-    // this.store.favorites.albums = this.store.favorites.albums.filter(
-    //   (itemId) => itemId !== id,
-    // );
-    // return `This action removes a #${id} album`;
+    // } catch (err) {
+    //   throw new HttpException(
+    //     {
+    //       state: StatusCodes.NOT_FOUND,
+    //       error: "record with id === userId doesn't exist",
+    //     },
+    //     StatusCodes.NOT_FOUND,
+    //   );
+    // }
   }
+
+  // this.store.tracks.forEach((track, index) => {
+  //   if (track.albumId === id) this.store.tracks[index].albumId = null;
+  // });
+
+  //
+  // const trackListToClean = await this.prisma.track.findMany({ where:  })
+
+  //
+  // this.store.favorites.albums = this.store.favorites.albums.filter(
+  //   (itemId) => itemId !== id,
+  // );
+  // return `This action removes a #${id} album`;
+  // }
 }
